+++ src/devices.json
