+++ src/new_core.py
@@ -14,9 +14,9 @@ class Main_utils_page(QWidget):
     def __init__(self):
         super().__init__()
         self.sig_use = pyqtSignal()
-        self.page = QWidget()
+        self.page = self.main_page()
+       
     def main_page(self):
-        self.ui = new_ui.Ui_MainWindow()
         page = QWidget()
         main_layout = QVBoxLayout()
         main_layout.setSpacing(0)
@@ -24,19 +24,16 @@ class Main_utils_page(QWidget):
         top_layout = QHBoxLayout()
         self.label = QLabel("Input Custom Words to Speak")
         top_layout.addWidget(self.label)
-
         bottom_half = QWidget()
         bottom_layout = QHBoxLayout() 
         self.label = QLabel("Input Custom Words to Speak")
         bottom_layout.addWidget(self.label)
-
         top_half.setLayout(top_layout)
         bottom_half.setLayout(bottom_layout)
         main_layout.addWidget(top_half,1)
         main_layout.addWidget(bottom_half,1)
         page.setLayout(main_layout)
-        self.ui.stackedWidget.addWidget(page)
-        self.ui.stackedWidget.setCurrentWidget(page)
         return page
+
