+++ src/styles.py
@@ -1,0 +1,49 @@
+# styles.py
+
+my_style = """
+QLineEdit {
+    border: 1px solid #ccc;
+    border-radius: 6px;
+    padding: 6px 10px;
+    background-color: #008a91;
+    color: #333;
+    font-size: 14px;
+}
+
+QLineEdit:focus {
+    border: 1px solid #0078d7;
+    background-color: #ffffff;
+}
+
+QLineEdit::placeholder {
+    color: #999;
+}
+
+QPushButton { 
+    padding: 5px 10px; 
+    border-radius: 5px; 
+    background-color: #4CAF50; 
+    color: white; 
+} 
+
+QLabel {
+    color: white;
+}
+"""
+
+btn_sheet = """
+QPushButton {
+    color: rgb(255, 255, 255);
+    background-color: #272757;   
+    border: 0px solid;
+    text-align: center;  
+    border-radius: 10px;
+}
+
+QPushButton:hover {
+    background-color: #9575cd;
+    font-size: 17px; 
+    border-radius: 10px;
+    margin: 5px;
+}
+"""
